backend/requirements.txt

flask
google-generativeai
asyncio
